use download job for dswxni
