+++ data_subscriber/query.py
@@ -109,7 +109,7 @@ class BaseQuery:
             job_submission_tasks = submit_rtc_download_job_submissions_tasks(batch_id_to_products_map.keys(), self.args, self.settings)
             results = asyncio.gather(*job_submission_tasks, return_exceptions=True)
         elif COLLECTION_TO_PRODUCT_TYPE_MAP[self.args.collection] == ProductType.NISAR_GCOV:
-            job_submision_tasks = self.submit_dswx_ni_job_submission_handler(download_granules, query_timerange)
+            job_submision_tasks = self.submit_gcov_download_job_submission_handler(download_granules, query_timerange)
             results = job_submision_tasks
         else:
             job_submission_tasks = self.download_job_submission_handler(download_granules, query_timerange)
