+++ docker/hysds-io.json.gcov_download
@@ -1,0 +1,50 @@
+{
+  "label": "Download GCOV data",
+  "submission_type":"individual",
+  "allowed_accounts": [ "ops" ],
+  "params": [
+    {
+      "name": "endpoint",
+      "from": "submitter",
+      "placeholder": "e.g. --endpoint=OPS",
+      "type": "text",
+      "default": "--endpoint=OPS"
+    },
+    {
+      "name": "batch_ids",
+      "from": "submitter",
+      "placeholder": "e.g. --batch-ids MS_1_1_15 MS_1_2_15",
+      "optional": true
+    },
+    {
+      "name": "smoke_run",
+      "from": "submitter",
+      "placeholder": "e.g. --smoke-run",
+      "optional": true
+    },
+    {
+      "name": "dry_run",
+      "from": "submitter",
+      "placeholder": "e.g. --dry-run",
+      "optional": true
+    },
+    {
+      "name": "transfer_protocol",
+      "from": "submitter",
+      "placeholder": "e.g. --transfer-protocol=auto",
+      "type": "text",
+      "default": "--transfer-protocol=auto",
+      "optional": true
+    },
+    {
+      "name": "dswx_ni_job_release",
+      "from": "submitter",
+      "placeholder": "e.g. --release-version=<release_version>"
+    },
+    {
+     "name": "product_metadata",
+     "from": "dataset_jpath:_source",
+     "lambda" : "lambda ds: { 'metadata': ds['metadata'] }"
+    }
+  ]
+}
